Decimalize scores
Before score value could be from -128 to 127, which is not too usable for tests which have decimal scores. Now it's changed and possible values are from -999.999 to 999.999. So now you can use scores like 0.005 or else. These changes are applied to scales's labels too.
Decimal-style values are shown only when it's needed. So if you have scale which values are 15 out of 15, it will be shown as "15 out of 15". But when values will be decimal — it will  be shown as decimal: "12.034 out of 24.3".

* WordPress compatibility updated to 4.2.4 and prepared to 4.3.
* Use [semantic headers](https://make.wordpress.org/core/2015/07/31/headings-in-admin-screens-change-in-wordpress-4-3/) in respondents results as of WP 4.3.

Fixes:

* Respect results orders. Helpful when you have many results and their output order is important at results page.
* Fix support tags generation.

Internal improvements:

* Step strategy know if answered questions are possible now.
* Step strategy can show step's description as a short description before questions.

git-svn-id: https://plugins.svn.wordpress.org/wp-testing/trunk@1216844 b8457f37-d9ea-0310-8a92-e5e31aec5664